requirements.txt
# Git & GitHub
PyGithub>=2.1.1
GitPython>=3.1.40

# Утилиты
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.7

ruff>=0.1.6
black>=23.11.0
mypy>=1.7.1
pytest>=7.4.3